# docs/openapi/permesi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "permesi",
    "description": "Identity and Access Management",
    "contact": {
      "name": "Team Permesi",
      "email": "team@permesi.dev"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.3.43"
  },
  "paths": {
...
    "/v1/auth/mfa/totp/enroll/finish": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Finish TOTP enrollment. Requires bootstrap or full session.",
        "operationId": "totp_enroll_finish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaTotpEnrollFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Enrollment finished"
          },
          "400": {
            "description": "Invalid code"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/auth/mfa/totp/enroll/start": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Start TOTP enrollment. Requires bootstrap or full session.",
        "operationId": "totp_enroll_start",
        "responses": {
          "200": {
            "description": "Enrollment started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaTotpEnrollStartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/v1/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "session",
        "responses": {
          "200": {
            "description": "Session is active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "204": {
            "description": "No active session"
          }
        }
      }
    },
...
    "/v1/me": {
      "get": {
        "tags": [
          "me"
        ],
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Return the authenticated user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session cookie."
          }
        }
      },
      "patch": {
        "tags": [
          "me"
        ],
        "operationId": "patch_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Profile updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid update payload."
          },
          "401": {
            "description": "Missing or invalid session cookie."
          }
        }
      }
    },
...
    "/v1/me/mfa/totp": {
      "delete": {
        "tags": [
          "me"
        ],
        "operationId": "disable_totp",
        "responses": {
          "204": {
            "description": "TOTP disabled."
          },
          "401": {
            "description": "Unauthorized or recent auth required."
          }
        }
      }
    },
    "/v1/me/mfa/webauthn/{credential_id}": {
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Deletes a registered `WebAuthn` security key.",
        "description": "Side Effects:\n- Disables MFA (`MfaState::Disabled`) if this was the last security key AND no TOTP is configured.\n- Logs an audit event.",
        "operationId": "delete_key",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "description": "Hex-encoded credential ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Security key deleted successfully"
          },
          "400": {
            "description": "Invalid credential id"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Security key not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MeResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "created_at",
          "updated_at",
          "roles",
          "scopes",
          "mfa_enabled",
          "totp_enabled"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "locale": {
            "type": [
              "string",
              "null"
            ]
          },
          "mfa_enabled": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totp_enabled": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "MeUpdateRequest": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "locale": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "MfaTotpEnrollFinishRequest": {
        "type": "object",
        "required": [
          "code",
          "credential_id"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "credential_id": {
            "type": "string"
          }
        }
      },
      "MfaTotpEnrollStartResponse": {
        "type": "object",
        "required": [
          "secret",
          "qr_code_url",
          "credential_id"
        ],
        "properties": {
          "credential_id": {
            "type": "string"
          },
          "qr_code_url": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        }
      },
...
      "SessionKind": {
        "type": "string",
        "description": "Session kinds used to gate MFA bootstrap and challenge flows.",
        "enum": [
          "full",
          "mfa_bootstrap",
          "mfa_challenge"
        ]
      },
      "SessionResponse": {
        "type": "object",
        "required": [
          "user_id",
          "email",
          "is_operator",
          "session_kind",
          "totp_enabled",
          "webauthn_enabled"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "is_operator": {
            "type": "boolean"
          },
          "session_kind": {
            "$ref": "#/components/schemas/SessionKind"
          },
          "totp_enabled": {
            "type": "boolean"
          },
          "user_id": {
            "type": "string"
          },
          "webauthn_enabled": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}